Move Dataset view to api_views Fixes #50
